requirements.txt
Django==6.0
asgiref==3.11.0
sqlparse==0.5.5
tzdata==2025.3

Pillow==12.0.0

channels==4.3.1
channels_redis==4.3.0
redis==7.0.1

cloudinary==1.44.1
requests==2.32.4

mysqlclient==2.2.7
